plot_agent/langraph_sql_agent/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-multipart==0.0.6
agno
clickhouse-driver
pandas
matplotlib
seaborn
streamlit==1.31.1
pillow
openai 
clickhouse_connect>=0.6.8
python-dotenv>=1.0.0
langgraph
langchain
langchain-openai
langchain-community
uuid